+++ Engine/Source/StateControl.cpp
@@ -112,7 +112,7 @@ void StateControl::SetString(wstring _key, const wstring& _value)
 	m_blackBoardString[_key] = _value;
 }
 
-void StateControl::AddInteger(wstring _key, float _addValue)
+void StateControl::AddInteger(wstring _key, int _addValue)
 {
 	m_blackBoardInteger[_key] += _addValue;
 }
@@ -121,8 +121,6 @@ void StateControl::AddFloat(wstring _key, float _addValue)
 {
 	m_blackBoardFloat[_key] += _addValue;
 }
-
-
 
 int StateControl::GetInteger(wstring _key)
 {
